phy: move regex pattern for parsing SimLogger logs to SimLogger class
